+++ format/processjson.go
@@ -16,11 +16,14 @@ package format
 
 import (
 	"encoding/json"
+	"github.com/mmcloughlin/geohash"
 	"github.com/mssola/user_agent"
 	"github.com/trivago/gollum/core"
 	"github.com/trivago/tgo/tcontainer"
 	"github.com/trivago/tgo/tmath"
 	"github.com/trivago/tgo/tstrings"
+	"gopkg.in/oschwald/geoip2-golang.v1"
+	"net"
 	"strconv"
 	"strings"
 	"time"
@@ -34,6 +37,7 @@ import (
 //  - "stream.Broadcast":
 //    Formatter: "format.ProcessJSON"
 //    ProcessJSONDataFormatter: "format.Forward"
+//    ProcessJSONGeoIPFile: ""
 //    ProcessJSONDirectives:
 //      - "host:split: :host:@timestamp"
 //      - "@timestamp:time:20060102150405:2006-01-02 15\\:04\\:05"
@@ -41,11 +45,18 @@ import (
 //      - "text:trim: \t"
 //      - "foo:rename:bar"
 //		- "foobar:remove"
+//      - "array:pick:0:firstOfArray"
+//		- "array:remove:foobar"
 //      - "user_agent:agent:browser:os:version"
+//      - "client:geoip:country:city:timezone:location"
 //    ProcessJSONTrimValues: true
 //
 // ProcessJSONDataFormatter formatter that will be applied before
 // ProcessJSONDirectives are processed.
+//
+// ProcessJSONGeoIPFile defines a GeoIP file to load. This enables the "geoip"
+// directive. If no file is loaded IPs will not be resolved. Files can be
+// found e.g. at http://dev.maxmind.com/geoip/geoip2/geolite2/
 //
 // ProcessJSONDirectives defines the action to be applied to the json payload.
 // Directives are processed in order of appearance.
@@ -57,17 +68,25 @@ import (
 //  * trim:<characters> remove the given characters (not string!) from the start
 //    and end of the value
 //  * rename:<old>:<new> rename a given field
-//  * remove remove a given field
+//  * remove remove a given field. If additional parameters are given, an array is
+//    expected. Strings given as additional parameters will be removed from that array
+//  * pick:<key>:<index>:<name> Pick a specific index from an array and store it
+//    in a new field.
 //  * time:<read>:<write> read a timestamp and transform it into another
 //    format
 //  * unixtimestamp:<read>:<write> read a unix timestamp and transform it into another
 //    format. valid read formats are s, ms, and ns.
-//  * agent:{<user_agent_field>:<user_agent_field>:...} Parse the value as a user
-//    agent string and extract the given fields into <key>_<user_agent_field>
-//    ("ua:agent:browser:os" would create the new fields "ua_browser" and "ua_os")
 //  * flatten:{<delimiter>} create new fields from the values in field, with new
 //    fields named field + delimiter + subfield. Delimiter defaults to ".".
 //    Removes the original field.
+//  * agent:<key>:{<field>:<field>:...} Parse the value as a user agent string and
+//    extract the given fields into <key>_<field>.
+//    ("ua:agent:browser:os" would create the new fields "ua_browser" and "ua_os").
+//    Possible values are: "mozilla", "platform", "os", "localization", "engine",
+//    "engine_version", "browser", "version".
+//  * geoip:{<field>:<field>:...} like agent this directive will analyse an IP string
+//    via geoip and produce new fields.
+//    Possible values are: "country", "city", "continent", "timezone", "proxy", "location"
 //
 // ProcessJSONTrimValues will trim whitspaces from all values if enabled.
 // Enabled by default.
@@ -75,6 +94,7 @@ type ProcessJSON struct {
 	core.SimpleFormatter
 	directives []transformDirective
 	trimValues bool
+	db         *geoip2.Reader
 }
 
 type transformDirective struct {
@@ -90,9 +110,22 @@ func init() {
 // Configure initializes this formatter with values from a plugin config.
 func (format *ProcessJSON) Configure(conf core.PluginConfigReader) error {
 	format.SimpleFormatter.Configure(conf)
+	if err != nil {
+		return err
+	}
+	directives := conf.GetStringArray("ProcessJSONDirectives", []string{})
 
-	format.trimValues = conf.GetBool("Trim", true)
-	directives := conf.GetStringArray("Directives", []string{})
+	format.base = plugin.(core.Formatter)
+	format.directives = make([]transformDirective, 0, len(directives))
+	format.trimValues = conf.GetBool("ProcessJSONTrimValues", true)
+	geoIPFile := conf.GetString("ProcessJSONGeoIPFile", "")
+
+	if geoIPFile != "" {
+		format.db, err = geoip2.Open(geoIPFile)
+		if err != nil {
+			return err
+		}
+	}
 
 	if len(directives) > 0 {
 		format.directives = make([]transformDirective, 0, len(directives))
@@ -123,7 +156,6 @@ func (format *ProcessJSON) Configure(conf core.PluginConfigReader) error {
 
 func (format *ProcessJSON) processDirective(directive transformDirective, values *tcontainer.MarshalMap) {
 	if value, keyExists := (*values)[directive.key]; keyExists {
-
 		numParameters := len(directive.parameters)
 		switch directive.operation {
 		case "rename":
@@ -189,9 +221,41 @@ func (format *ProcessJSON) processDirective(directive transformDirective, values
 				}
 			}
 
+		case "pick":
+			index, _ := strconv.Atoi(directive.parameters[0])
+			field := directive.parameters[1]
+			array, _ := (*values).Array(directive.key)
+
+			if index < 0 {
+				index = len(array) - index
+			}
+			if index < 0 || index >= len(array) {
+				Log.Warning.Print("Array index out of bounds")
+				return
+			}
+			(*values)[field] = array[index]
+
 		case "remove":
 			if numParameters == 0 {
 				delete(*values, directive.key)
+			} else {
+				array, _ := (*values).Array(directive.key)
+				wi := 0
+			nextItem:
+				for ri, value := range array {
+					if strValue, isStr := value.(string); isStr {
+						for _, cmp := range directive.parameters {
+							if strValue == cmp {
+								continue nextItem
+							}
+						}
+					}
+					if ri != wi {
+						array[wi] = array[ri]
+					}
+					wi++
+				}
+				(*values)[directive.key] = array[:wi]
 			}
 
 		case "replace":
@@ -216,6 +280,26 @@ func (format *ProcessJSON) processDirective(directive transformDirective, values
 			case numParameters == 1:
 				(*values)[directive.key] = strings.Trim(stringValue, tstrings.Unescape(directive.parameters[0]))
 			}
+
+		case "flatten":
+			delimiter := "."
+			if numParameters == 1 {
+				delimiter = directive.parameters[0]
+			}
+			keyPrefix := directive.key + delimiter
+			if mapValue, err := values.Map(directive.key); err == nil {
+				for key, val := range mapValue {
+					(*values)[keyPrefix+key.(string)] = val
+				}
+			} else if arrayValue, err := values.Array(directive.key); err == nil {
+				for index, val := range arrayValue {
+					(*values)[keyPrefix+strconv.Itoa(index)] = val
+				}
+			} else {
+				Log.Warning.Print("key was not a JSON array or object: " + directive.key)
+				return
+			}
+			delete(*values, directive.key)
 
 		case "agent":
 			stringValue, err := values.String(directive.key)
@@ -258,25 +342,71 @@ func (format *ProcessJSON) processDirective(directive transformDirective, values
 				}
 			}
 
-		case "flatten":
-			delimiter := "."
-			if numParameters == 1 {
-				delimiter = directive.parameters[0]
-			}
-			keyPrefix := directive.key + delimiter
-			if mapValue, err := values.Map(directive.key); err == nil {
-				for key, val := range mapValue {
-					(*values)[keyPrefix + key.(string)] = val
-				}
-			} else if arrayValue, err := values.Array(directive.key); err == nil {
-				for index, val := range arrayValue {
-					(*values)[keyPrefix + strconv.Itoa(index)] = val
-				}
-			} else {
-				format.Log.Warning.Print("key was not a JSON array or object: " + directive.key)
+		case "geoip":
+			if format.db == nil {
 				return
 			}
-			delete(*values, directive.key)
+
+			ipString, err := (*values).String(directive.key)
+			if err != nil {
+				Log.Warning.Print(err.Error())
+				return
+			}
+
+			fields := []string{
+				"country",
+				"city",
+				"continent",
+				"timezone",
+				"proxy",
+				"location",
+				}
+			if numParameters > 0 {
+				fields = directive.parameters
+				}
+
+			ip := net.ParseIP(ipString)
+			record, err := format.db.City(ip)
+			if err != nil {
+				Log.Warning.Printf("IP \"%s\" could not be resolved: %s", ipString, err.Error())
+				Log.Debug.Printf("%#v", values)
+				return
+			}
+
+			for _, field := range fields {
+				switch field {
+				case "city":
+					name, exists := record.City.Names["en"]
+					if exists {
+						(*values)[directive.key+"_city"] = name
+					}
+
+				case "country":
+					(*values)[directive.key+"_countryCode"] = record.Country.IsoCode
+					name, exists := record.Country.Names["en"]
+					if exists {
+						(*values)[directive.key+"_country"] = name
+			}
+
+				case "continent":
+					(*values)[directive.key+"_continentCode"] = record.Continent.Code
+					name, exists := record.Continent.Names["en"]
+					if exists {
+						(*values)[directive.key+"_continent"] = name
+					}
+
+				case "timezone":
+					(*values)[directive.key+"_timezone"] = record.Location.TimeZone
+
+				case "proxy":
+					(*values)[directive.key+"_proxy"] = record.Traits.IsAnonymousProxy
+					(*values)[directive.key+"_satellite"] = record.Traits.IsSatelliteProvider
+
+				case "location":
+					(*values)[directive.key+"_geocoord"] = []float64{record.Location.Latitude, record.Location.Longitude}
+					(*values)[directive.key+"_geohash"] = geohash.Encode(record.Location.Latitude, record.Location.Longitude)
+				}
+			}
 		}
 	}
 }