+++ CHANGELOG.md
@@ -12,6 +12,10 @@ This is a patch / minor features release.
  * Consumer.Kafka does now retry with an "oldest" offset after encountering an OutOfRange exception.
  * Fixed a crash when using producer.ElasticSearch with date based indexes (thanks @relud)
  * format.Base64Decode now uses data from previous formatters as intended
+ * format.JSON arr and obj will now auto create a key if necessary
+ * format.JSON now checks for valid state references upon startup
+ * format.JSON now properly encodes strings when using "enc"
+ * format.SplitToJSON may now keep JSON payload and is better at escaping string
  
 #### New
 
@@ -22,10 +26,16 @@ This is a patch / minor features release.
  * Added authentication support to consumer.Kafka (thanks @relud)
  * Added consumer group support to consumer.Kafka (thanks @relud)
  * Added a native SystemD consumer (thanks @relud)
+ * Added a Statsd producer for counting messages (thanks @relud)
  * Added an option to flatten JSON arrays into single values with format.ProcessJSON (thanks @relud)
  * Added filter.Any to allow "or" style combinations of filters (thanks @relud)
  * Added support for unix timestamp parsing to format.ProcessJSON (thanks @relud)
  * Added filter.Sample to allow processing of every n'th message only (thanks @relud)
+ * Added format.TemplateJSON to apply golang templates to JSON payloads (thanks @relud)
+ * Added "pick" option to format.ProcessJSON to get a single value from an arrays
+ * Extended "remove" option pf format.ProcessJSON to remove values from arrays
+ * Added "geoip" option to format.ProcessJSON to get GeoIP data from an IP
+ * Added index configuration options to producer.ElasticSearch
 
 # 0.4.4